feat(mini-vue): createApp train
