+++ core/index.js
@@ -1,1 +1,2 @@
 export { Dep, effectWatch, reactive } from "./reactivity.js";
+export { createApp } from "./createApp.js";
